requirements-client.txt
python-rtmidi==0.5b1
requests==2.6.0
socketIO-client==0.6.1
